Kdv Hesaplama

+++ src/Task_2_KDV.java
@@ -11,7 +11,7 @@ public class Task_2_KDV {
 
 
         System.out.println("Ürün tutarını giriniz:"); // ürün tutarı alınır
-        t= input.nextInt();
+        t= input.nextDouble();
 
         if (t<=1000){ // Koşulları tanımlama 1000'den büyük ise %8 değil ise %18
             kdv=t*0.18;
